visioncaptioning/requirements.txt
Pillow==10.0.0
fastapi == 0.99.1
uvicorn == 0.23.2
python-multipart == 0.0.6
timm==0.4.12
transformers==4.31.0
fairscale==0.4.4
pycocoevalcap